pyproject.toml
[project]
name = "autograb"
version = "0.1.0"
description = "A simple coding interview project"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "pytest>=7.0.0",
    "pytest-sugar>=0.9.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "ipython>=8.12.3",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.8"



[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--color=yes"
]
pythonpath = ["src"]

[tool.pytest-sugar]
verbose = true

[tool.autograb.database]
host = "localhost"
port = 5432
database = "autograb"
username = "autograb_user"
password = "autograb_password" 
